Working on implementing firebase, sign in with google. Need to find out a way to assosciate the access token from spotify with a user id, (probably can send a code to the client then use that in a subsequent api request) but I should finish the firebase and sign in with google implementation first.
